pyproject.toml
[project]
name = "emailassistant2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "grandalf>=0.8",
    "ipython>=9.8.0",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.0",
    "langgraph>=1.0.5",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
